+++ src/main/java/org/scoula/controller/EmailController.java
@@ -2,7 +2,6 @@ package org.scoula.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.scoula.security.dto.AuthDTO;
 import org.scoula.security.dto.MemberDTO;
 import org.scoula.service.EmailService;
 import org.scoula.service.EmailVerificationService;
@@ -16,17 +15,17 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/v1/signup")
+@RequestMapping("/v1/email")
 @RequiredArgsConstructor
 @Log4j2
-public class SignUpController {
+public class EmailController {
     private final EmailService emailService;
     private final EmailVerificationService emailVerification;
     private final UserService userService;
 
 
 
-    @PostMapping("/email")
+    @PostMapping("")
     public ResponseEntity<?> sendVerificationCode(@RequestBody Map<String, String> request) {
         String email = request.get("user_id");
         emailService.sendVerificationCode(email);
@@ -35,7 +34,7 @@ public class SignUpController {
         return ResponseEntity.ok(Map.of("message","검증 코드 발송 완료!"));
     }
 
-    @PostMapping("/email/verification")
+    @PostMapping("/verification")
     public ResponseEntity<?> verifyEmailCode(@RequestBody Map<String, String> request) {
         String email = request.get("user_id");
         String inputCode = request.get("code");
@@ -50,16 +49,7 @@ public class SignUpController {
             return ResponseEntity.badRequest().body(Map.of("message", "이메일 인증 실패"));
         }
     }
-    @PostMapping("")
-    public ResponseEntity<?> signUp(@RequestBody MemberDTO memberDTO) {
-        try {
-            userService.signUp(memberDTO); // ✅ 트랜잭션 처리된 메소드 호출
-            return ResponseEntity.ok(Map.of("message", "회원가입 완료"));
-        } catch (Exception e) {
-            log.error("[signUp] 회원가입 실패", e);
-            return ResponseEntity.internalServerError().body(Map.of("message", "회원가입 실패"));
-        }
-    }
+
